edited configure.ac to add command-line option to configure script for Altera SDK. Programs run on the FPGA must be pre-processed with different headers and linked with different libraries so it is necessary to specify the target device at configure time. this is accomplished with the commands:
% cd OpenDwarfs
% ./autogen.sh
% mkdir fpga-build && cd fpga-build
% ../configure --with-AOCL=/path/to/AOCL

the configure script now defines the macro USE_AFPGA when the Altera SDK
is specified in the call to configure as above. This can and should be
used in the host code because the setup process is a little different
for Altera devices than cpu/gpus. Note that Altera has not upgraded from
OpenCL 1.0

if the --with-AOCL option is not passed to configure, the macro USEGPU
will be defined, and all dwarves have been updated to reflect this. 

These macros are defined in config.h which must be included in the
source.

csr.c has been updated to take advantage of this compile-time macro. The
command-line switch for using the FPGA has been eliminated and the
run-time if statements have been replaced with compile-time ifs. 

SPMV has been tested successfully on intel CPU with intel SDK as well as
Nallatech board with Altera FPGA & SDK for small (N=4) inputs. Seg fault
has been observed on FPGA for large (N=131072) inputs.