removed redefinition of CHKERR from several dwarfs, this macro should be included from common_ocl.h
changed csr.c to use 3 command-queues (one for writing,executing, and
reading) to pipeline the kernel executions of each matrix in input file
with pcie read & writes for previous and next kernels. 

The global memory buffers are all allocated immediately after reading
the input file and not released until the very end of the program, so
the total memory allocated still cannot exceed 4GB.

According to altera's programming guide, it should be possible to
enqueue at least 1000 kernels, but we are getting a crash on the 501st
enqueue. Input files with less than 500 matrices seem to run without
problems.